requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
pymongo
certifi
pyyaml
dill

# -e .